requirements.txt
openai-whisper
ffmpeg>=1.4
mysql-connector-python>=8.0.32
flask>=2.2.3